+++ tests/resources/modules/Sample/Livewire/HelloTable.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Modules\Sample\Livewire;
+
+use Livewire\Component;
+
+class HelloTable extends Component
+{
+    public function render()
+    {
+        return view('sample::hello')->layout('layout::admin');
+    }
+}