+++ tests/resources/modules/Sample/Views/hello.blade.php
@@ -1,0 +1,1 @@
+<div>HELLO FROM SAMPLE MODULE</div>