+++ tests/resources/modules/Sample/Views/menu.blade.php
@@ -1,0 +1,1 @@
+<x-rpd::nav-link label="Sample" route="sample.hello" />