+++ tests/resources/modules/Sample/config.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    'layout' => 'layout::admin',
+    'menu_admin' => 'sample::menu',
+    'menu_admin_position' => 99,
+];